Hadoop compatibility now also supports MR2 API.
